+++ src/AngleSharp.Diffing/Core/AttributeComparison.cs
@@ -1,3 +1,5 @@
+using System.Runtime.InteropServices;
+
 namespace AngleSharp.Diffing.Core;
 
 /// <summary>
@@ -5,11 +7,12 @@ namespace AngleSharp.Diffing.Core;
 /// </summary>
 /// <param name="Control">Gets the control attribute which the <see cref="Test"/> attribute is supposed to match.</param>
 /// <param name="Test">Gets the test attribute which should be compared to the <see cref="Control"/> attribute.</param>
+[StructLayout(LayoutKind.Auto)]
 public readonly record struct AttributeComparison(in AttributeComparisonSource Control, in AttributeComparisonSource Test)
 {
     /// <summary>
     /// Returns the control and test elements which the control and test attributes belongs to.
     /// </summary>
-    public (IElement ControlElement, IElement TestElement) GetAttributeElements()
+    public readonly (IElement ControlElement, IElement TestElement) AttributeElements
         => ((IElement)Control.ElementSource.Node, (IElement)Test.ElementSource.Node);
 }