+++ src/AngleSharp.Diffing/Core/CompareResult.cs
@@ -10,39 +10,44 @@ public readonly record struct CompareResult(CompareDecision Decision, IDiff? Dif
     /// <summary>
     /// Use when the compare result is unknown.
     /// </summary>
-    public static readonly CompareResult Unknown = default;
+    public static readonly CompareResult Unknown;
 
     /// <summary>
     /// Use when the two compared nodes or attributes are the same.
     /// </summary>
-    public static readonly CompareResult Same = new CompareResult(CompareDecision.Same);
+    public static readonly CompareResult Same = new(CompareDecision.Same);
 
     /// <summary>
     /// Use when the comparison should be skipped and any child-nodes or attributes skipped as well.
     /// </summary>
-    public static readonly CompareResult Skip = new CompareResult(CompareDecision.Skip);
+    public static readonly CompareResult Skip = new(CompareDecision.Skip);
 
     /// <summary>
     /// Use when the comparison should skip any child-nodes.
     /// </summary>
-    public static readonly CompareResult SkipChildren = new CompareResult(CompareDecision.SkipChildren);
+    public static readonly CompareResult SkipChildren = new(CompareDecision.SkipChildren);
 
     /// <summary>
     /// Use when the comparison should skip any attributes.
     /// </summary>
-    public static readonly CompareResult SkipAttributes = new CompareResult(CompareDecision.SkipAttributes);
+    public static readonly CompareResult SkipAttributes = new(CompareDecision.SkipAttributes);
+
+    /// <summary>
+    /// Use when the comparison should skip any attributes.
+    /// </summary>
+    public static readonly CompareResult SkipChildrenAndAttributes = new(CompareDecision.SkipChildren | CompareDecision.SkipAttributes);
 
     /// <summary>
     /// Use when the two compared nodes or attributes are the different.
     /// </summary>
-    public static CompareResult Different => new CompareResult(CompareDecision.Different);
+    public static CompareResult Different => new(CompareDecision.Different);
 
     /// <summary>
     /// Use when the two compared nodes or attributes are the different.
     /// </summary>
     /// <param name="diff">The associated <see cref="IDiff"/> describing the difference.</param>
     /// <returns>Returns a <see cref="CompareResult"/> with <see cref="CompareResult.Decision"/> set to <see cref="CompareDecision.Different"/>.</returns>
-    public static CompareResult FromDiff(IDiff diff) => new CompareResult(CompareDecision.Different, diff);
+    public static CompareResult FromDiff(IDiff diff) => new(CompareDecision.Different, diff);
 
     /// <summary>
     /// Checks if a <see cref="CompareResult"/> is either a <see cref="CompareResult.Same"/> or <see cref="CompareResult.Skip"/>.