+++ src/AngleSharp.Diffing/Extensions/EmptyHtmlCollection.cs
@@ -8,10 +8,12 @@ public class EmptyHtmlCollection<T> : IHtmlCollection<T> where T : IElement
 {
     /// <inheritdoc/>
     [SuppressMessage("Design", "CA1065:Do not raise exceptions in unexpected locations")]
+    [SuppressMessage("Usage", "CA2201:Do not raise reserved exception types")]
     public T this[int index] => throw new IndexOutOfRangeException();
 
     /// <inheritdoc/>
     [SuppressMessage("Design", "CA1065:Do not raise exceptions in unexpected locations")]
+    [SuppressMessage("Usage", "CA2201:Do not raise reserved exception types")]
     public T this[string id] => throw new IndexOutOfRangeException();
 
     /// <inheritdoc/>