+++ src/AngleSharp.Diffing/Strategies/AttributeStrategies/ClassAttributeComparer.cs
@@ -18,11 +18,11 @@ public static class ClassAttributeComparer
         if (!IsBothClassAttributes(comparison))
             return CompareResult.FromDiff(new AttrDiff(comparison, AttrDiffKind.Name));
 
-        var (ctrlElm, testElm) = comparison.GetAttributeElements();
+        var (ctrlElm, testElm) = comparison.AttributeElements;
         if (ctrlElm.ClassList.Length != testElm.ClassList.Length)
             return CompareResult.FromDiff(new AttrDiff(comparison, AttrDiffKind.Value));
 
-        return ctrlElm.ClassList.All(x => testElm.ClassList.Contains(x))
+        return ctrlElm.ClassList.All(x => testElm.ClassList.Contains(x, StringComparer.Ordinal))
             ? CompareResult.Same
             : CompareResult.FromDiff(new AttrDiff(comparison, AttrDiffKind.Value));
     }